# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API Playlist Vidoe",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/bookreturns": {
            "get": {
                "description": "list bookreturn entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List bookreturn entities",
                "operationId": "list-bookreturn",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Bookreturn"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create bookreturn",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create bookreturn",
                "operationId": "create-bookreturn",
                "parameters": [
                    {
                        "description": "bookreturn entity",
                        "name": "bookreturn",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Bookreturn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Bookreturn"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/bookreturns/{id}": {
            "delete": {
                "description": "get bookreturn by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a bookreturn entity by ID",
                "operationId": "delete-bookreturn",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bookreturn ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/books": {
            "get": {
                "description": "list book entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List book entities",
                "operationId": "list-book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Book"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create book",
                "operationId": "create-book",
                "parameters": [
                    {
                        "description": "Book entity",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Book"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Book"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "description": "get book by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a book entity by ID",
                "operationId": "get-book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Book"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update book by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a book entity by ID",
                "operationId": "update-book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Book entity",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Book"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Book"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get book by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a book entity by ID",
                "operationId": "delete-book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.Book": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "integer"
                },
                "bookname": {
                    "type": "string"
                },
                "category": {
                    "type": "integer"
                },
                "userid": {
                    "type": "integer"
                }
            }
        },
...
        "controllers.Bookreturn": {
            "type": "object",
            "properties": {
                "bookborrowID": {
                    "type": "integer"
                },
                "locationID": {
                    "type": "integer"
                },
                "returnTime": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
...
        "ent.Author": {
            "type": "object",
            "properties": {
                "Name": {
                    "description": "Name holds the value of the \"Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the AuthorQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.AuthorEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.AuthorEdges": {
            "type": "object",
            "properties": {
                "owner": {
                    "description": "Owner holds the value of the owner edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Research"
                    }
                },
                "writer": {
                    "description": "Writer holds the value of the writer edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Book"
                    }
                }
            }
        },
        "ent.Book": {
            "type": "object",
            "properties": {
                "BookName": {
                    "description": "BookName holds the value of the \"BookName\" field.",
                    "type": "string"
                },
                "author_ID": {
                    "type": "integer"
                },
                "category_id": {
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BookQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BookEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "status_ID": {
                    "type": "integer"
                },
                "user_ID": {
                    "type": "integer"
                }
            }
        },
        "ent.BookEdges": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "Author holds the value of the author edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Author"
                },
                "booklist": {
                    "description": "Booklist holds the value of the booklist edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Bookborrow"
                    }
                },
                "category": {
                    "description": "Category holds the value of the category edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Category"
                },
                "status": {
                    "description": "Status holds the value of the status edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Status"
                },
                "user": {
                    "description": "User holds the value of the user edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
        "ent.Bookborrow": {
            "type": "object",
            "properties": {
                "BORROW_DATE": {
                    "description": "BORROWDATE holds the value of the \"BORROW_DATE\" field.",
                    "type": "string"
                },
                "RETURN_DATE": {
                    "description": "RETURNDATE holds the value of the \"RETURN_DATE\" field.",
                    "type": "string"
                },
                "book_ID": {
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BookborrowQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BookborrowEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "servicepoint_ID": {
                    "type": "integer"
                },
                "user_ID": {
                    "type": "integer"
                }
            }
        },
        "ent.BookborrowEdges": {
            "type": "object",
            "properties": {
                "book": {
                    "description": "BOOK holds the value of the BOOK edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Book"
                },
                "borrowed": {
                    "description": "Borrowed holds the value of the borrowed edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Bookreturn"
                    }
                },
                "servicepoint": {
                    "description": "SERVICEPOINT holds the value of the SERVICEPOINT edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ServicePoint"
                },
                "user": {
                    "description": "USER holds the value of the USER edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
...
        "ent.Bookreturn": {
            "type": "object",
            "properties": {
                "RETURN_TIME": {
                    "description": "RETURNTIME holds the value of the \"RETURN_TIME\" field.",
                    "type": "string"
                },
                "client_ID": {
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BookreturnQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BookreturnEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "location_ID": {
                    "type": "integer"
                },
                "user_ID": {
                    "type": "integer"
                }
            }
        },
        "ent.BookreturnEdges": {
            "type": "object",
            "properties": {
                "location": {
                    "description": "Location holds the value of the location edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Location"
                },
                "mustreturn": {
                    "description": "Mustreturn holds the value of the mustreturn edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Bookborrow"
                },
                "user": {
                    "description": "User holds the value of the user edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
        "ent.Category": {
            "type": "object",
            "properties": {
                "CategoryName": {
                    "description": "CategoryName holds the value of the \"CategoryName\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the CategoryQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.CategoryEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.CategoryEdges": {
            "type": "object",
            "properties": {
                "catof": {
                    "description": "Catof holds the value of the catof edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Book"
                    }
                }
            }
        },
...
        "ent.Location": {
            "type": "object",
            "properties": {
                "LOCATION_NAME": {
                    "description": "LOCATIONNAME holds the value of the \"LOCATION_NAME\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the LocationQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.LocationEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.LocationEdges": {
            "type": "object",
            "properties": {
                "returnfrom": {
                    "description": "Returnfrom holds the value of the returnfrom edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Bookreturn"
                    }
                }
            }
        },
...
        "ent.Status": {
            "type": "object",
            "properties": {
                "STATUS_NAME": {
                    "description": "STATUSNAME holds the value of the \"STATUS_NAME\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the StatusQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.StatusEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.StatusEdges": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "Status holds the value of the status edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.ClientEntity"
                    }
                },
                "statusofbook": {
                    "description": "Statusofbook holds the value of the statusofbook edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Book"
                    }
                }
            }
        },
        "ent.User": {
            "type": "object",
            "properties": {
                "PASSWORD": {
                    "description": "PASSWORD holds the value of the \"PASSWORD\" field.",
                    "type": "string"
                },
                "USER_EMAIL": {
                    "description": "USEREMAIL holds the value of the \"USER_EMAIL\" field.",
                    "type": "string"
                },
                "USER_NAME": {
                    "description": "USERNAME holds the value of the \"USER_NAME\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UserQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UserEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "role_ID": {
                    "type": "integer"
                }
            }
        },
        "ent.UserEdges": {
            "type": "object",
            "properties": {
                "addby": {
                    "description": "Addby holds the value of the addby edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Book"
                    }
                },
                "booking": {
                    "description": "Booking holds the value of the booking edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Booking"
                    }
                },
                "borrow": {
                    "description": "Borrow holds the value of the borrow edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Bookborrow"
                    }
                },
                "position": {
                    "description": "Position holds the value of the position edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Role"
                },
                "preemption": {
                    "description": "Preemption holds the value of the preemption edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Preemption"
                    }
                },
                "record": {
                    "description": "Record holds the value of the record edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Research"
                    }
                },
                "return": {
                    "description": "Return holds the value of the return edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Bookreturn"
                    }
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}